Added wrapped io.reader/writer parser/writer
